Add function to check if pokemon is at selected location
This moves the call of checkLocations to within the new function markAsFound,
which is passed as a prop to the Selection component.

checkLocations returns a promise, the resolved value of which is true or false.
If the returned value is true and the pokemon has not already been found,
then the pokemon is marked as found by updating the state object
and the relevant styles are applied to it.